+++ src/core/sockimpl.h
@@ -105,7 +105,7 @@ struct subinfo {
 };
 
 struct nni_pipe {
-	uint32_t           p_id;
+	uint64_t           p_id;
 	nni_sp_pipe_ops    p_tran_ops;
 	nni_proto_pipe_ops p_proto_ops;
 	size_t             p_size;
